complete day-5 tuples

+++ day_5/27_list_max_min.py
@@ -1,5 +1,5 @@
 num_list = input('Enter number list: ').split()
-num_list = [int(n) for n in num_list] # convert elements to int for proper comparison
+num_list = [float(n) for n in num_list] # convert elements to int for proper comparison
 
 min_num = num_list[0]
 max_num = num_list[0]
